// pages/types.ts
export type WarehouseType = 'finished' | 'parts' | 'raw' | 'catering';

export interface Product {
  id: string;
  name: string;
  barcode: string;
  jdeCode?: string;
  jdeCodeBulk?: string;
  jdeCodePacked?: string;
  price: number;
  cost: number;
  stock: number;
  stockBulk?: number;
  stockPacked?: number;
  initialStockBulk?: number; 
  initialStockPacked?: number; 
  category: string;
  image?: string;
  warehouse: WarehouseType;
  unit?: string;
  minStock?: number;
  maxStock?: number;
  reorderPoint?: number; 
  sackWeight?: number;
  supplierId?: string;
  customFields?: Record<string, string>;
  goodsGroup?: string; 
  housingOfficer?: string;
  notes?: string;
}